itfMesManager: long->int fixes for FUN_003a2900 (diff 57->47) and FUN_003a2970 (diff 127->40); FUN_0035f140 returns s32

+++ src/itfMesManager.c
@@ -829,7 +829,7 @@ u32 FUN_003a2900(void)
 
 {
 
-  long lVar1;
+  int lVar1;
 
   
 
@@ -861,7 +861,7 @@ u8 FUN_003a2970(void)
 
   u8 bVar2;
 
-  long lVar3;
+  int lVar3;
 
   u32 uVar4;
 
@@ -877,7 +877,7 @@ u8 FUN_003a2970(void)
 
   else {
 
-    puVar1 = *(u32 **)(DAT_00959eec_abs + (int)lVar3 * 0xd);
+    puVar1 = *(u32 **)(DAT_00959eec_abs + lVar3 * 0xd);
 
     if (puVar1 == (u32 *)0x0) {
